+++ src/utils.rs
@@ -2,6 +2,7 @@ use byteorder::{BigEndian, ByteOrder, LittleEndian};
 use std::io::{self, Read, Write};
 
 use crate::block::Block;
+use crate::Float;
 
 // ---------------------------------------------------------------------------
 // GCD helpers
@@ -47,13 +48,13 @@ pub fn compute_min_gcd(blocks: &[Block]) -> usize {
 
 /// Component-wise subtraction.
 #[inline]
-pub(crate) fn sub3(a: [f64; 3], b: [f64; 3]) -> [f64; 3] {
+pub(crate) fn sub3(a: [Float; 3], b: [Float; 3]) -> [Float; 3] {
     [a[0] - b[0], a[1] - b[1], a[2] - b[2]]
 }
 
 /// Cross product.
 #[inline]
-pub(crate) fn cross3(a: [f64; 3], b: [f64; 3]) -> [f64; 3] {
+pub(crate) fn cross3(a: [Float; 3], b: [Float; 3]) -> [Float; 3] {
     [
         a[1] * b[2] - a[2] * b[1],
         a[2] * b[0] - a[0] * b[2],
@@ -63,25 +64,25 @@ pub(crate) fn cross3(a: [f64; 3], b: [f64; 3]) -> [f64; 3] {
 
 /// Dot product.
 #[inline]
-pub(crate) fn dot3(a: [f64; 3], b: [f64; 3]) -> f64 {
+pub(crate) fn dot3(a: [Float; 3], b: [Float; 3]) -> Float {
     a[0] * b[0] + a[1] * b[1] + a[2] * b[2]
 }
 
 /// Euclidean norm.
 #[inline]
-pub(crate) fn vec_norm3(a: [f64; 3]) -> f64 {
+pub(crate) fn vec_norm3(a: [Float; 3]) -> Float {
     dot3(a, a).sqrt()
 }
 
 /// Euclidean distance between two 3-D points.
 #[inline]
-pub(crate) fn distance3(a: [f64; 3], b: [f64; 3]) -> f64 {
+pub(crate) fn distance3(a: [Float; 3], b: [Float; 3]) -> Float {
     vec_norm3(sub3(a, b))
 }
 
 /// Apply a 3×3 rotation matrix to a point.
 #[inline]
-pub(crate) fn apply_rotation(p: [f64; 3], rot: [[f64; 3]; 3]) -> [f64; 3] {
+pub(crate) fn apply_rotation(p: [Float; 3], rot: [[Float; 3]; 3]) -> [Float; 3] {
     [
         rot[0][0] * p[0] + rot[0][1] * p[1] + rot[0][2] * p[2],
         rot[1][0] * p[0] + rot[1][1] * p[1] + rot[1][2] * p[2],
